Route GET and POST /subcategories
