+++ api/src/routes/subCategories.js
@@ -1,0 +1,64 @@
+const { Router } = require('express');
+const { grape } = require('../db')
+const { SubCategory } = require('../db')
+const { Category } = require('../db')
+
+const router = Router();
+
+//Con esta pequeña funcion, seteo las primeras letras de los string a mayusculas
+// const firstUpperCase = function (mayus) { return mayus.replace(/\b\w/g, l => l.toUpperCase()) }
+
+try {
+  router.get('/', async (req, res) => {
+    let { type } = req.query
+    if (type) {
+      const subCategoryDB = await SubCategory.findOne({ where: { type } })
+      if (subCategoryDB) {
+        res.json(subCategoryDB)
+      } else {
+        res.status(404).json(`La subcategoria ${type} es inexistente en este e-commerce\nComuniquese con la gente del backend`)
+      }
+    } else {
+      const allsubcategorias = await SubCategory.findAll()
+      res.json(allsubcategorias)
+    }
+  })
+} catch (error) {
+  console.log(error)
+  res.status(404).json(error)
+}
+
+
+try {
+  router.post('/', async(req, res) => {
+    const { type, categoryId } = req.body
+ 
+    const newSubCategory = await SubCategory.create({type})
+    newSubCategory.setCategory(categoryId)
+    res.json(newSubCategory)
+  })
+} catch (error) {
+  console.log(error)
+  res.status(404).json(error)
+}
+// router.get('/grape', async (req, res) => {
+// console.log("esto es lo que hay en db grape:", grape)
+//     let { name } = req.query
+//     if (name) {
+//         name = firstUpperCase(name)
+//         const uva = await grape.findAll({
+//             where: { name }
+//         })
+//         return (uva) ? res.json(uva)
+//             : res.status(404).json(`No tenemos existencias en nuestro catalogo del tipo de uva que estas buscando`)
+//     }
+//     const alluvas = await grape.findAll()
+//     res.json(alluvas)
+// })
+
+// router.get('/products', async (req, res) => {
+// res.json(`esperese mijito que esta ruta aun no funciona`)
+// })
+
+
+module.exports = router;
